+++ core/src/main/java/org/mule/runtime/core/api/store/PartitionableExpirableObjectStore.java
@@ -11,5 +11,5 @@ import java.io.Serializable;
 public interface PartitionableExpirableObjectStore<T extends Serializable>
     extends ExpirableObjectStore<T>, PartitionableObjectStore<T> {
 
-  void expire(int entryTTL, int maxEntries, String partitionName) throws ObjectStoreException;
+  void expire(long entryTTL, int maxEntries, String partitionName) throws ObjectStoreException;
 }
